IMMORTAL: Add definitions.h and move enums str, motives, and cyc from story.h to definitions.h
